vector_indexer/pyproject.toml
[project]
name = "vector-indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "orjson>=3.11.4",
    "pika==1.3.2",
    "pydantic-settings==2.10.0",
    "weaviate-client>=4.17.0",
]
